finish the config function and download function
